libft/lambdas/finance/arroweye-sftp/requirements.txt
python-gnupg
pysftp
boto3
requests
jsonpickle
aws-xray-sdk
cachetools
